Refactor scoreboard state management with Context API
