+++ src/components/App.js
@@ -36,39 +36,6 @@ const App = () => {
     setHighScore(Math.max(...scores));
   }, [players]);
 
-  const handleRemovePlayer = (id) => {
-    setPlayers((prevPlayers) =>
-      prevPlayers.filter((p) => p.id !== id)
-    );
-  };
-
-  const handleScoreChange = (id, delta) => {
-    setPlayers((prevPlayers) =>
-      prevPlayers.map((player) => {
-        if (player.id === id) {
-          return {
-            name: player.name,
-            score: player.score + delta,
-            id: player.id,
-          };
-        }
-        return player;
-      })
-    );
-  };
-
-  const handleAddPlayer = (name) => {
-    setPlayers((prevPlayers) => [
-      ...prevPlayers,
-      {
-        name,
-        score: 0,
-        id: nextPlayerId.current,
-      },
-    ]);
-    nextPlayerId.current += 1;
-  };
-
   return (
     <div className="scoreboard">
       <Header players={players} />
