+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './components/App';
+import App from './Components/App';
 import './index.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
